alterado a função dinamic screen agora só realiza o plot na transição, uma variável teve mudança de tipo

+++ tela.h
@@ -44,7 +44,7 @@ extern "C"{
     typedef struct {  
         uint8_t h_peep;
         uint8_t h_pressao;
-        uint8_t h_prop;
+        uint16_t h_prop;
         uint16_t h_volume;
         uint16_t h_freq;
         uint16_t h_temp_insp;
@@ -55,7 +55,6 @@ extern "C"{
     void screen_static(char p);           //tela Static
     void screen_dynamic(config_t *IHM_aux, char p, uint8_t cursor); 
     void screen_Init();
-
     void set_IHM_volume(config_t *IHM_aux, uint8_t p);
     void set_IHM_peep(config_t *IHM_aux, uint8_t p);
     void set_IHM_pressao(config_t *IHM_aux, uint8_t p);
